parakeet-stt-daemon/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "parakeet-stt-daemon"
version = "0.1.0"
description = "Local Parakeet push-to-talk daemon (WebSocket API + NeMo streaming)."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.115.5",
  "uvicorn[standard]>=0.32.1",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.1",
  "websockets>=13.1",
  "numpy>=2.1.3",
  "sounddevice>=0.4.7",
  "loguru>=0.7.2",
]

[project.optional-dependencies]
inference = [
  "torch==2.4.1",
  "torchaudio==2.4.1",
  "nemo_toolkit[asr]==2.0.0rc0",
]

[project.scripts]
parakeet-stt-daemon = "parakeet_stt_daemon.__main__:main"

[dependency-groups]
dev = [
  "ruff>=0.7.0",
  "black>=24.10.0",
  "pyright>=1.1.389",
]

[tool.hatch.build.targets.wheel]
packages = ["src/parakeet_stt_daemon"]

[tool.hatch.build.targets.sdist]
include = ["src/parakeet_stt_daemon", "README.md", "pyproject.toml"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [".venv", ".uv"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W", "UP"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
reportMissingImports = "warning"
